Minimum required JBoss modules to provide a HotRod server
This amounts to almost 10MB, most of which is Scala, in which HotRod
is written. Due to the fact that the "endpoints" module pulls in not
only HotRod but memcached, rest, and websockets, we're getting quite a
bit more than we want. We'll defer that problem for now.

It's worth mentioning that the
infinispan-server-infinispan-5.3.0.Final.jar includes the
org.jboss.as.clustering.infinispan classes we're replacing in the
corresponding module.xml.